requirements.txt
griptape>=0.21.1
openai>=1.12.0
anthropic>=0.8.1
groq>=0.4.1
google-generativeai>=0.3.2
cohere>=4.43
emergence>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
sympy>=1.12
matplotlib>=3.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic-core>=2.14.0
torch>=2.1.0
transformers>=4.36.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0
